# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS API",
        "description": "Learning Management System API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/",
            "description": "LMS API"
        }
    ],
    "paths": {
        "/intake/course": {
            "post": {
                "tags": [
                    "Intakes"
                ],
                "summary": "Add missing intake unit",
                "description": "Add missing intake unit",
                "operationId": "intakeCourse",
                "parameters": [
                    {
                        "name": "intake_course_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "unit_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "401": {
                        "description": "Unathorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}